fix projectaa: distinguish different logs
`RaftLog.entries` are the in-memory logs and logs in `RaftLog.storage` are
the logs store in storage but are not compacted (the first log is
compacted, which is used to track the prev index). So in projecta, I
only need the logs in `RaftLog.entries`. For the further projects, I need
retrieve the first index and the last index of the current storage when
I call `newLog`. And I also need to retrieve the entries from storage
into memory. Every operation of log index should be based on the first
log index of current `RaftLog.storage`. The changes are:

- log: `newLog` first retrieve entries from storage and record the first
index log index.
- log: every operation relevant to index is based on `RaftLog.firstIndex`.
- storage: adding `Append` method to `Storage` interface in the last commit is wrong, remove it.